requirements.txt
requests
python-dotenv
langgraph
langchain-core
langchain-community
langchain-openai
requests